Add bookings

+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -2,13 +2,14 @@ package ru.practicum.shareit.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 
 @Data
 @AllArgsConstructor
 public class UserDto {
-    private int id;
+    private long id;
     @NotBlank
     private String name;
     @Email
